requirements.txt
accelerate==0.30.1
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
astrapy==1.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bson==0.5.10
cassandra-driver==3.29.1
cassio==0.1.7
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.6
Deprecated==1.2.14
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
filelock==3.14.0
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.5.0
geomet==0.2.1.post1
gradientai==1.13.0
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
hyperframe==6.0.1
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
llama-index==0.10.42
llama-index-agent-openai==0.2.6
llama-index-cli==0.1.12
llama-index-core==0.10.42
llama-index-embeddings-gradient==0.1.2
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-gradient==0.1.2
llama-index-llms-huggingface==0.2.2
llama-index-llms-openai==0.1.21
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.23
llama-index-readers-llama-parse==0.1.4
llama-index-vector-stores-cassandra==0.1.3
llama-parse==0.4.4
llamaindex-py-client==0.1.19
MarkupSafe==2.1.5
marshmallow==3.21.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openai==1.30.5
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psutil==5.9.8
pydantic==2.7.2
pydantic_core==2.18.3
pypdf==4.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
striprtf==0.0.26
sympy==1.12.1
tbb==2021.12.0
tenacity==8.3.0
text-generation==0.7.0
tiktoken==0.4.0
tokenizers==0.19.1
toml==0.10.2
torch==2.3.0
tqdm==4.66.4
transformers==4.41.2
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
uuid6==2024.1.12
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
